optionally add docker host to proxy excludes in env command
This commit adds --no-proxy flag to the env command.
This will take care of adding the docker host to the NO_PROXY environment variable.

This change helps using docker-machine with docker >=1.5 with local vm providers (e.g. fusion, virtualbox) in coporate networks where an http proxy is required for internet connectivity.
Since version 1.5 the docker client picks up the HTTP_PROXY environment variable when connecting to the docker daemon. This does not work for the local vm usecase as the docker host is only reachble from the local workstation.

Signed-off-by: Fabian Ruff <fabian@progra.de>